+++ forui/lib/src/foundation/portal/portal_shift.dart
@@ -3,93 +3,93 @@ import 'package:flutter/widgets.dart';
 
 import 'package:forui/src/foundation/rendering.dart';
 
-/// A portal's target.
-typedef FPortalTarget = ({Offset offset, Size size, Alignment anchor});
+/// A portal child's rectangle.
+typedef FPortalChildBox = ({Offset offset, Size size, Alignment anchor});
 
-/// A portal's follower.
-typedef FPortalFollower = ({Size size, Alignment anchor});
+/// A portal's rectangle.
+typedef FPortalBox = ({Size size, Alignment anchor});
 
-/// Provides various follower shifting strategies for when a follower overflows out of the viewport.
-extension FPortalFollowerShift on Never {
-  /// Flips the follower to the opposite side of the target if it does not cause the follower to overflow out of the
-  /// viewport. Otherwise shifts the follower along the target's edge.
-  static Offset flip(Size view, FPortalTarget target, FPortalFollower follower) {
-    var anchor = none(view, target, follower).translate(target.offset.dx, target.offset.dy);
+/// Provides various portal shifting strategies for when a portal overflows out of the viewport.
+extension FPortalShift on Never {
+  /// Flips the portal to the opposite side of the child if it does not cause the portal to overflow out of the
+  /// viewport. Otherwise shifts the portal along the child's edge.
+  static Offset flip(Size view, FPortalChildBox child, FPortalBox portal) {
+    var anchor = none(view, child, portal).translate(child.offset.dx, child.offset.dy);
 
     final viewBox = Offset.zero & view;
-    final followerBox = anchor & follower.size;
+    final portalBox = anchor & portal.size;
 
-    switch ((viewBox, followerBox)) {
-      case _ when followerBox.left < viewBox.left:
-        final flipped = _flip(target, follower, x: true);
-        if ((flipped & follower.size).right <= viewBox.right) {
+    switch ((viewBox, portalBox)) {
+      case _ when portalBox.left < viewBox.left:
+        final flipped = _flip(child, portal, x: true);
+        if ((flipped & portal.size).right <= viewBox.right) {
           anchor = flipped;
         }
 
-      case _ when viewBox.right < followerBox.right:
-        final flipped = _flip(target, follower, x: true);
-        if (viewBox.left <= (flipped & follower.size).left) {
+      case _ when viewBox.right < portalBox.right:
+        final flipped = _flip(child, portal, x: true);
+        if (viewBox.left <= (flipped & portal.size).left) {
           anchor = flipped;
         }
     }
 
-    switch ((viewBox, followerBox)) {
-      case _ when followerBox.top < viewBox.top:
-        final flipped = _flip(target, follower, x: false);
-        if ((flipped & follower.size).bottom <= viewBox.bottom) {
+    switch ((viewBox, portalBox)) {
+      case _ when portalBox.top < viewBox.top:
+        final flipped = _flip(child, portal, x: false);
+        if ((flipped & portal.size).bottom <= viewBox.bottom) {
           anchor = flipped;
         }
 
-      case _ when viewBox.bottom < followerBox.bottom:
-        final flipped = _flip(target, follower, x: false);
-        if (viewBox.top <= (flipped & follower.size).top) {
+      case _ when viewBox.bottom < portalBox.bottom:
+        final flipped = _flip(child, portal, x: false);
+        if (viewBox.top <= (flipped & portal.size).top) {
           anchor = flipped;
         }
     }
 
-    final foo = _along(anchor, viewBox, anchor & follower.size);
-    return foo.translate(-target.offset.dx, -target.offset.dy);
+    final foo = _along(anchor, viewBox, anchor & portal.size);
+    return foo.translate(-child.offset.dx, -child.offset.dy);
   }
 
-  static Offset _flip(FPortalTarget target, FPortalFollower follower, {required bool x}) {
-    final targetAnchor = x ? target.anchor.flipX() : target.anchor.flipY();
-    final followerAnchor = x ? follower.anchor.flipX() : follower.anchor.flipY();
+  static Offset _flip(FPortalChildBox child, FPortalBox portal, {required bool x}) {
+    final childAnchor = x ? child.anchor.flipX() : child.anchor.flipY();
+    final portalAnchor = x ? portal.anchor.flipX() : portal.anchor.flipY();
 
-    final anchor = targetAnchor.relative(to: target.size) - followerAnchor.relative(to: follower.size);
+    final anchor = childAnchor.relative(to: child.size) - portalAnchor.relative(to: portal.size);
 
-    return anchor.translate(target.offset.dx, target.offset.dy);
+    return anchor.translate(child.offset.dx, child.offset.dy);
   }
 
-  /// Shifts the follower along the target's edge if the follower overflows out of the viewport.
-  static Offset along(Size view, FPortalTarget target, FPortalFollower follower) {
-    final anchor = none(view, target, follower).translate(target.offset.dx, target.offset.dy);
+  /// Shifts the portal along the child's edge if the portal overflows out of the viewport.
+  static Offset along(Size view, FPortalChildBox child, FPortalBox portal) {
+    final anchor = none(view, child, portal).translate(child.offset.dx, child.offset.dy);
 
     final viewBox = Offset.zero & view;
-    final followerBox = anchor & follower.size;
+    final portalBox = anchor & portal.size;
 
-    return _along(anchor, viewBox, followerBox).translate(-target.offset.dx, -target.offset.dy);
+    return _along(anchor, viewBox, portalBox).translate(-child.offset.dx, -child.offset.dy);
   }
 
-  static Offset _along(Offset anchor, Rect viewBox, Rect followerBox) {
-    anchor = switch ((viewBox, followerBox)) {
-      _ when followerBox.left < viewBox.left => Offset(anchor.dx + (viewBox.left - followerBox.left), anchor.dy),
-      _ when viewBox.right < followerBox.right => Offset(anchor.dx - followerBox.right + viewBox.right, anchor.dy),
+  static Offset _along(Offset anchor, Rect viewBox, Rect portalBox) {
+    anchor = switch ((viewBox, portalBox)) {
+      _ when portalBox.left < viewBox.left => Offset(anchor.dx + (viewBox.left - portalBox.left), anchor.dy),
+      _ when viewBox.right < portalBox.right => Offset(anchor.dx - portalBox.right + viewBox.right, anchor.dy),
       _ => anchor,
     };
 
-    anchor = switch ((viewBox, followerBox)) {
-      _ when followerBox.top < viewBox.top => Offset(anchor.dx, anchor.dy + (viewBox.top - followerBox.top)),
-      _ when viewBox.bottom < followerBox.bottom => Offset(anchor.dx, anchor.dy - followerBox.bottom + viewBox.bottom),
+    anchor = switch ((viewBox, portalBox)) {
+      _ when portalBox.top < viewBox.top => Offset(anchor.dx, anchor.dy + (viewBox.top - portalBox.top)),
+      _ when viewBox.bottom < portalBox.bottom => Offset(anchor.dx, anchor.dy - portalBox.bottom + viewBox.bottom),
       _ => anchor,
     };
 
     return anchor;
   }
 
-  /// Does not perform any shifting if the follower overflows out of the viewport.
-  static Offset none(Size view, FPortalTarget target, FPortalFollower follower) {
-    final targetAnchor = target.anchor.relative(to: target.size);
-    final followerAnchor = follower.anchor.relative(to: follower.size);
-    return targetAnchor - followerAnchor;
+  /// Does not perform any shifting if the portal overflows out of the viewport.
+  static Offset none(Size view, FPortalChildBox child, FPortalBox portal) {
+    final childAnchor = child.anchor.relative(to: child.size);
+    final portalAnchor = portal.anchor.relative(to: portal.size);
+    return childAnchor - portalAnchor;
   }
 }