+++ forui/lib/src/widgets/select_menu_tile.dart
@@ -56,10 +56,10 @@ class FSelectMenuTile<T> extends FormField<Set<T>> with FTileMixin {
   final Alignment tileAnchor;
 
   /// The shifting strategy used to shift a menu when it overflows out of the viewport. Defaults to
-  /// [FPortalFollowerShift.flip].
+  /// [FPortalShift.flip].
   ///
-  /// See [FPortalFollowerShift] for more information on the different shifting strategies.
-  final Offset Function(Size, FPortalTarget, FPortalFollower) shift;
+  /// See [FPortalShift] for more information on the different shifting strategies.
+  final Offset Function(Size, FPortalChildBox, FPortalBox) shift;
 
   /// True if the popover is hidden when tapped outside of it. Defaults to true.
   final bool hideOnTapOutside;
@@ -124,7 +124,7 @@ class FSelectMenuTile<T> extends FormField<Set<T>> with FTileMixin {
     this.divider = FTileDivider.full,
     this.menuAnchor = Alignment.topRight,
     this.tileAnchor = Alignment.bottomRight,
-    this.shift = FPortalFollowerShift.flip,
+    this.shift = FPortalShift.flip,
     this.hideOnTapOutside = true,
     this.directionPadding = false,
     this.autoHide = false,
@@ -255,7 +255,7 @@ class FSelectMenuTile<T> extends FormField<Set<T>> with FTileMixin {
     this.divider = FTileDivider.full,
     this.menuAnchor = Alignment.topRight,
     this.tileAnchor = Alignment.bottomRight,
-    this.shift = FPortalFollowerShift.flip,
+    this.shift = FPortalShift.flip,
     this.hideOnTapOutside = true,
     this.directionPadding = false,
     this.autoHide = false,